+++ nglview/parameters.py
@@ -1,4 +1,4 @@
-REPR_NAME_PAIRS = [
+REPRESENTATION_NAME_PAIRS = [
         ('axes', 'axes'),
         ('point', 'point'),
         ('line', 'line'),
@@ -21,4 +21,4 @@ REPR_NAME_PAIRS = [
         ('ball_and_stick', 'ball+stick'),
         ]
 
-REPR_NAMES = list(sorted(set([name for pairs in REPR_NAME_PAIRS for name in pairs])))
+REPRESENTATION_NAMES = list(sorted(set([name for pairs in REPRESENTATION_NAME_PAIRS for name in pairs])))
